backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Дмитрий Спиряев <imwisagist@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = {extras = ["standart"], version = "^0.21.1"}
requests = "^2.28.2"
SQLAlchemy = "^2.0.9"
psycopg2 = "^2.9.6"
alembic = "^1.10.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
